requirements.txt
# Core web framework and ASGI server
fastapi
uvicorn[standard]

# LangChain ecosystem
langchain
langchain-openai
langchain-community  # For additional document loaders
langgraph

# Database and ORM
sqlmodel
sqlalchemy
psycopg2-binary  # PostgreSQL adapter (optional)

# Document processing
docling

# Vector database
chromadb
langchain-chroma

# Text processing and utilities
tiktoken
numpy

# Additional dependencies for document handling
pydantic
python-dotenv
PyJWT
python-multipart

# UI and API client
streamlit
requests

# Web content fetching
firecrawl-py  # Professional web scraping service
beautifulsoup4  # HTML parsing
lxml  # XML/HTML parser
aiohttp  # Async HTTP client